rectifier simulation
formed a comprehensive simulation focusing on the rectifier part of the circuit. the placeholder buck converter is placed to model a more realistic demand from the load. there is also a starter resistor and bypass circuitry that is not connected to the main model. it will probably be removed after further analysis and comparison when the remaining elements of the circuit are completed as well.